my lab6 assignment
this is mineeee

+++ Source/AssignmentLab6/src/com/assignment/lab/HeightChecker.java
@@ -2,9 +2,9 @@ package com.assignment.lab;
 
 public class HeightChecker {
 
-public int heightc(double feet, double inch) {
-	int feet2 = (int) (feet * 30.48);
-	int inch2 = (int) (inch * 2.54);
+public double heightc(double feet, double inch) {
+	double feet2 =  (feet * 30.48);
+	double inch2 =  (inch * 2.54);
 	System.out.println("values: " +feet2  + " + " + inch2);
 	return feet2 + inch2;
 }
